% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AleStrategy.R
\name{AleStrategy}
\alias{AleStrategy}
\title{AleStrategy: Generalized Additive Decomposition Based on ALE Effects}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{EffectStrategy}.
}
\description{
ALE-based effect strategy (inherits from \link{EffectStrategy}). Given model and data,
preprocesses to Z/Y via \code{prepare_split_data_ale};
transforms ALE effects per node; computes ALE-derivative heterogeneity;
finds best split via \code{search_best_split_ale}; fits tree and plots ALE curves.
}
\details{
Intended for use through \code{GadgetTree$new(strategy = AleStrategy$new())} and
\code{tree$fit(...)}. Can be instantiated directly for custom pipelines.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{s = AleStrategy$new(categorical_split = "ordered_prefix")
}\if{html}{\out{</div>}}
}

\examples{
ale_strat = AleStrategy$new()
ale_strat$categorical_split

}
\seealso{
\code{\link{calculate_ale_heterogeneity_cpp}}
}
\section{Super class}{
\code{::EffectStrategy} -> \code{AleStrategy}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{(\code{any}) \cr
Fitted model (persistent after \code{$fit()}).}

\item{\code{data}}{(\code{data.frame()} or \code{data.table()}) \cr
Data (persistent after \code{$fit()}).}

\item{\code{target_feature_name}}{(\code{character(1)}) \cr
Target variable name.}

\item{\code{n_intervals}}{(\code{integer(1)}) \cr
Intervals for numeric ALE.}

\item{\code{predict_fun}}{(\verb{function()}) \cr
\code{function(model, data)} returning predictions.}

\item{\code{order_method}}{(\code{character(1)}) \cr
Categorical order: \code{"mds"}, \code{"pca"}, \code{"random"}, \code{"raw"}.}

\item{\code{ale_engine}}{(\code{character(1)}) \cr
ALE backend selected after \code{$fit()}: \code{"cpp"} or \code{"r"}.}

\item{\code{categorical_split}}{(\code{character(1)}) \cr
Categorical split mode for ALE trees: \code{"ordered_prefix"} or \code{"exhaustive"}.}

\item{\code{max_exhaustive_levels}}{(\code{integer(1)}) \cr
Maximum observed levels allowed for exhaustive categorical split search.}

\item{\code{effect}}{(\code{list()} or \code{NULL}) \cr
Cached ALE effect used when \code{$plot()} omits \code{effect}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AleStrategy-new}{\code{AleStrategy$new()}}
\item \href{#method-AleStrategy-preprocess}{\code{AleStrategy$preprocess()}}
\item \href{#method-AleStrategy-node_transform}{\code{AleStrategy$node_transform()}}
\item \href{#method-AleStrategy-heterogeneity}{\code{AleStrategy$heterogeneity()}}
\item \href{#method-AleStrategy-get_child_objectives}{\code{AleStrategy$get_child_objectives()}}
\item \href{#method-AleStrategy-find_best_split}{\code{AleStrategy$find_best_split()}}
\item \href{#method-AleStrategy-plot}{\code{AleStrategy$plot()}}
\item \href{#method-AleStrategy-fit}{\code{AleStrategy$fit()}}
\item \href{#method-AleStrategy-clean}{\code{AleStrategy$clean()}}
\item \href{#method-AleStrategy-clone}{\code{AleStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AleStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-AleStrategy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AleStrategy$new(
  categorical_split = "ordered_prefix",
  max_exhaustive_levels = 12L
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AleStrategy-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-AleStrategy-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AleStrategy$preprocess(
  model,
  effect = NULL,
  data,
  target_feature_name,
  n_intervals,
  feature_set = NULL,
  split_feature = NULL,
  predict_fun = NULL,
  order_method = "raw",
  ale_engine = c("auto", "cpp", "r")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AleStrategy-node_transform"></a>}}
\if{latex}{\out{\hypertarget{method-AleStrategy-node_transform}{}}}
\subsection{Method \code{node_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AleStrategy$node_transform(Y, idx, grid = NULL, is_child = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AleStrategy-heterogeneity"></a>}}
\if{latex}{\out{\hypertarget{method-AleStrategy-heterogeneity}{}}}
\subsection{Method \code{heterogeneity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AleStrategy$heterogeneity(Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AleStrategy-get_child_objectives"></a>}}
\if{latex}{\out{\hypertarget{method-AleStrategy-get_child_objectives}{}}}
\subsection{Method \code{get_child_objectives()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AleStrategy$get_child_objectives(
  Y,
  split_info,
  idx_left,
  idx_right,
  grid_left,
  grid_right
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AleStrategy-find_best_split"></a>}}
\if{latex}{\out{\hypertarget{method-AleStrategy-find_best_split}{}}}
\subsection{Method \code{find_best_split()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AleStrategy$find_best_split(Z, Y, min_node_size, n_quantiles)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AleStrategy-plot"></a>}}
\if{latex}{\out{\hypertarget{method-AleStrategy-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AleStrategy$plot(
  tree,
  effect = NULL,
  data,
  target_feature_name,
  depth = NULL,
  node_id = NULL,
  features = NULL,
  show_plot = TRUE,
  show_point = TRUE,
  mean_center = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AleStrategy-fit"></a>}}
\if{latex}{\out{\hypertarget{method-AleStrategy-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AleStrategy$fit(
  tree,
  model = NULL,
  effect = NULL,
  data,
  target_feature_name,
  n_intervals = 10,
  feature_set = NULL,
  split_feature = NULL,
  predict_fun = NULL,
  order_method = "raw",
  ale_engine = c("auto", "cpp", "r"),
  categorical_split = NULL,
  max_exhaustive_levels = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AleStrategy-clean"></a>}}
\if{latex}{\out{\hypertarget{method-AleStrategy-clean}{}}}
\subsection{Method \code{clean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AleStrategy$clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AleStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AleStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AleStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
