% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ale_heterogeneity.R
\name{calculate_ale_heterogeneity_cpp}
\alias{calculate_ale_heterogeneity_cpp}
\title{Calculate ALE Heterogeneity}
\usage{
calculate_ale_heterogeneity_cpp(Y)
}
\arguments{
\item{Y}{(\code{list()} or \code{data.frame()}) \cr
ALE effect data.}
}
\value{
(\code{numeric(1)} or \code{list()}) \cr
Heterogeneity value(s): single value when Y is data.frame, list with one value per feature when Y is list.
}
\description{
Calculate ALE Heterogeneity
}
\keyword{internal}
