% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{create_plots_for_depth}
\alias{create_plots_for_depth}
\title{Create PD plots for one depth.}
\usage{
create_plots_for_depth(
  tree,
  prepared_data,
  data,
  target_feature_name,
  depth_idx,
  nodes_to_render,
  color_ice,
  color_pd,
  show_plot,
  show_point,
  mean_center
)
}
\arguments{
\item{tree}{(\code{list()}) \cr
Depth-based list of Node objects.}

\item{prepared_data}{(\code{list()}) \cr
Prepared effect data per feature.}

\item{data}{(\code{data.frame()} or \code{data.table()}) \cr
Original data.}

\item{target_feature_name}{(\code{character(1)}) \cr
Target column name.}

\item{depth_idx}{(\code{integer(1)}) \cr
Depth level.}

\item{nodes_to_render}{(\code{integer()}) \cr
Node indices to render.}

\item{color_ice, color_pd}{(\code{character(1)}) \cr
Colors for ICE and PDP.}

\item{show_plot, show_point, mean_center}{(\code{logical(1)}) \cr
Plot options.}
}
\value{
(\code{list()}) \cr
Named list of patchwork objects per node.
}
\description{
Create PD plots for one depth.
}
\keyword{internal}
