% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data_ale.R
\name{cumsum_na_as_zero}
\alias{cumsum_na_as_zero}
\title{Cumulative sum with NA as zero.}
\usage{
cumsum_na_as_zero(values)
}
\arguments{
\item{values}{(\code{numeric()}) \cr
Values to cumulatively sum; \code{NA} treated as 0.}
}
\value{
(\code{numeric()}) \cr
Cumulative sum.
}
\description{
Cumulative sum with NA as zero.
}
\keyword{internal}
