% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_split_data_utils.R
\name{ensure_factors}
\alias{ensure_factors}
\title{Ensure character columns are factors.}
\usage{
ensure_factors(data, cols)
}
\arguments{
\item{data}{(\code{data.frame()} or \code{data.table()}) \cr
Data to modify.}

\item{cols}{(\code{character()}) \cr
Column names to ensure as factor.}
}
\value{
(\code{data.frame()} or \code{data.table()}) \cr
Copy of \code{data} with the listed character columns converted to factor.
}
\description{
Ensure character columns are factors.
}
\keyword{internal}
