% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_to_numeric.R
\name{factor_to_numeric}
\alias{factor_to_numeric}
\title{Convert factor to numeric.}
\usage{
factor_to_numeric(f)
}
\arguments{
\item{f}{(\code{factor()}) \cr
Factor to convert.}
}
\value{
(\code{numeric()}) \cr
Numeric vector.
}
\description{
Given f: if all level labels parse as numbers,
uses as.numeric(as.character(f)); otherwise uses as.numeric(f) (level indices).
Returns numeric vector.
}
\keyword{internal}
