% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PdStrategy.R
\name{PdStrategy}
\alias{PdStrategy}
\title{PdStrategy: Generalized Additive Decomposition Based on PD Effects}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{EffectStrategy}.
}
\description{
PD-based effect strategy (inherits from \link{EffectStrategy}). Given effect or model and data,
preprocesses to Z/Y/grid;
mean-centers effects per node; computes sum-of-variances heterogeneity;
finds best split via C++; fits tree and plots PD/ICE.
Character feature columns are coerced to \code{factor} before ICE/PD computation so they match
split-matrix treatment and learner conventions (same as \code{prepare_split_data_common}).
}
\details{
This class is used internally by the GadgetTree framework to implement partial dependence
tree growing, splitting, and visualization. It is not intended to be used directly by end users,
but can be instantiated for advanced customization.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{s = PdStrategy$new(categorical_split = "one_vs_rest")
}\if{html}{\out{</div>}}
}

\examples{
pd_strat = PdStrategy$new()
pd_strat$categorical_split

}
\section{Super class}{
\code{::EffectStrategy} -> \code{PdStrategy}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{effect}}{(\code{list()} or \code{R6} or \code{NULL}) \cr
Cached PD/ICE effect used when \code{$plot()} omits \code{effect}.}

\item{\code{categorical_split}}{(\code{character(1)}) \cr
Categorical split mode for PD trees: \code{"one_vs_rest"} or \code{"exhaustive"}.}

\item{\code{max_exhaustive_levels}}{(\code{integer(1)}) \cr
Maximum observed levels allowed for exhaustive categorical split search.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PdStrategy-new}{\code{PdStrategy$new()}}
\item \href{#method-PdStrategy-preprocess}{\code{PdStrategy$preprocess()}}
\item \href{#method-PdStrategy-node_transform}{\code{PdStrategy$node_transform()}}
\item \href{#method-PdStrategy-heterogeneity}{\code{PdStrategy$heterogeneity()}}
\item \href{#method-PdStrategy-get_child_objectives}{\code{PdStrategy$get_child_objectives()}}
\item \href{#method-PdStrategy-find_best_split}{\code{PdStrategy$find_best_split()}}
\item \href{#method-PdStrategy-plot}{\code{PdStrategy$plot()}}
\item \href{#method-PdStrategy-fit}{\code{PdStrategy$fit()}}
\item \href{#method-PdStrategy-clean}{\code{PdStrategy$clean()}}
\item \href{#method-PdStrategy-clone}{\code{PdStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-PdStrategy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdStrategy$new(categorical_split = "one_vs_rest", max_exhaustive_levels = 12L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdStrategy-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-PdStrategy-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdStrategy$preprocess(
  effect,
  data,
  target_feature_name = NULL,
  feature_set = NULL,
  split_feature = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdStrategy-node_transform"></a>}}
\if{latex}{\out{\hypertarget{method-PdStrategy-node_transform}{}}}
\subsection{Method \code{node_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdStrategy$node_transform(Y, idx, grid, is_child = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdStrategy-heterogeneity"></a>}}
\if{latex}{\out{\hypertarget{method-PdStrategy-heterogeneity}{}}}
\subsection{Method \code{heterogeneity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdStrategy$heterogeneity(Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdStrategy-get_child_objectives"></a>}}
\if{latex}{\out{\hypertarget{method-PdStrategy-get_child_objectives}{}}}
\subsection{Method \code{get_child_objectives()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdStrategy$get_child_objectives(
  Y,
  split_info,
  idx_left,
  idx_right,
  grid_left,
  grid_right
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdStrategy-find_best_split"></a>}}
\if{latex}{\out{\hypertarget{method-PdStrategy-find_best_split}{}}}
\subsection{Method \code{find_best_split()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdStrategy$find_best_split(Z, Y, min_node_size, n_quantiles)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdStrategy-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PdStrategy-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdStrategy$plot(
  tree,
  effect = NULL,
  data,
  target_feature_name,
  depth = NULL,
  node_id = NULL,
  features = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdStrategy-fit"></a>}}
\if{latex}{\out{\hypertarget{method-PdStrategy-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdStrategy$fit(
  tree,
  effect = NULL,
  model = NULL,
  data,
  target_feature_name,
  feature_set = NULL,
  split_feature = NULL,
  predict_fun = NULL,
  n_grid = 20L,
  pd_engine = c("auto", "cpp", "r"),
  categorical_split = NULL,
  max_exhaustive_levels = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdStrategy-clean"></a>}}
\if{latex}{\out{\hypertarget{method-PdStrategy-clean}{}}}
\subsection{Method \code{clean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdStrategy$clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PdStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
