% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{select_depths_to_render}
\alias{select_depths_to_render}
\title{Select depth indices to render based on tree, depth, and node_id.}
\usage{
select_depths_to_render(tree, depth, node_id)
}
\arguments{
\item{tree}{(\code{list()}) \cr
Depth-based list of Node objects.}

\item{depth}{(\code{integer()} or \code{NULL}) \cr
Depth levels to consider; \code{NULL} = all.}

\item{node_id}{(\code{integer()} or \code{NULL}) \cr
Node IDs to render; \code{NULL} = all nodes at selected depths.}
}
\value{
(\code{integer()}) \cr
Depth indices to render.
}
\description{
Select depth indices to render based on tree, depth, and node_id.
}
\keyword{internal}
