% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplaineff_internal.R
\name{xplaineff_internal}
\alias{xplaineff_internal}
\alias{ale_sweep_cpp}
\alias{calculate_ale_matrix}
\alias{calculate_ale_heterogeneity_list_cpp}
\alias{calculate_ale_heterogeneity_single_cpp}
\alias{re_mean_center_ice_cpp}
\alias{search_best_split_cpp}
\alias{default_predict_fun}
\alias{make_effect_predictor}
\alias{predict_newdata_fast_dispatch}
\alias{fast_predict_regression_model}
\alias{predict_ranger_regression_fast}
\alias{extract_ranger_regression_model}
\alias{extract_ranger_num_threads}
\alias{predict_rpart_regression_fast}
\alias{extract_rpart_regression_model}
\alias{predict_xgboost_regression_fast}
\alias{extract_xgboost_regression_model}
\alias{xgboost_booster_is_regression}
\alias{extract_mlr3_native_model}
\alias{extract_mlr3_feature_names}
\alias{select_newdata_features}
\alias{numeric_matrix_for_prediction}
\alias{extract_numeric_prediction}
\alias{has_predict_method}
\alias{cpp_pd_stack_newdata}
\alias{risk_from_stats}
\alias{assert_ale_effect_list}
\alias{active_effect_rel_tol}
\alias{active_effect_indices}
\alias{active_effect_names}
\alias{subset_ale_compact_features}
\alias{prune_effects_for_split_search}
\alias{d_l}
\alias{interval_index}
\alias{level}
\alias{x}
\alias{x_grid}
\alias{x_left}
\alias{x_right}
\alias{y}
\title{Internal C++ helpers and package symbols}
\description{
Functions and symbols used internally by the package. Not intended for direct use.
}
\keyword{internal}
